reuse extract function as exist check
